chore: separate method to trait
